flask/requirements.txt
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
greenlet==3.0.3
Pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5
mariadb==1.1.10
gunicorn==20.0.4
# mahnungsgenerator
docxtpl==0.19.1
qrbill==1.1.0
svglib==1.5.1
CairoSVG==2.7.1